Move current implementation out of `current` folder
